+++ CHANGELOG.md
@@ -2,20 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### 0.1.3 (2023-03-14)
+### 0.1.4 (2023-03-15)
 
 
 ### Features
 
 * **controller:** added controller, display and adjusted network slightly ([c633736](https://github.com/xjonsson/bitburner-scripts/commit/c6337369958f92d2c12bf45de07032fcafde4934))
+* **controller:** added working version of batch controller ([123d0fc](https://github.com/xjonsson/bitburner-scripts/commit/123d0fcd881cf5efbe492068bfd345b87e17ea43))
 * **controller:** merging this in so people can use it ([9ac1040](https://github.com/xjonsson/bitburner-scripts/commit/9ac10409dfb27e37c9a46cfd9d5b4aecd48b550c))
+* **display:** added ability to show targets and their value ([3f543b1](https://github.com/xjonsson/bitburner-scripts/commit/3f543b17288abffc6937266265bb96e7c0e2b43a))
+* **display:** added display for network and statistics ([5b4e9cc](https://github.com/xjonsson/bitburner-scripts/commit/5b4e9ccd45802e12e863c02c2a5e34de1c0886eb))
 * **minideploy:** added script for minideploy from home ([13025c9](https://github.com/xjonsson/bitburner-scripts/commit/13025c9652122b3ee763b6e244a6eadbcdd694ba))
 * **minimal:** added base for minimal ([28e0414](https://github.com/xjonsson/bitburner-scripts/commit/28e04141c1084fea7afe178a11eca5f7a5180e93))
 * **minimal:** added focus target and display ([99f9dbd](https://github.com/xjonsson/bitburner-scripts/commit/99f9dbdf529670e5453a524b868c0fe896e90c94))
 * **minimal:** added reclaim function to hack servers ([abab0d9](https://github.com/xjonsson/bitburner-scripts/commit/abab0d939dc400a556a35f08495a3fe57338cbf3))
 * **minimal:** added xmin for bot usage and working version of minimal ([9eda81d](https://github.com/xjonsson/bitburner-scripts/commit/9eda81de7db94fd4acf201c19a65fb0dab6ff42b))
 * **network:** added networking class to handle servers, bots, and ring ([cd340d2](https://github.com/xjonsson/bitburner-scripts/commit/cd340d2ea0e51144791c0c84c46ff55431c08311))
+* **network:** added ring count to networking ([ee77b83](https://github.com/xjonsson/bitburner-scripts/commit/ee77b83b70fb8230e6cfc8e314adb7225d9fea18))
 * **player:** added basic helper class to help with player data ([d3521e7](https://github.com/xjonsson/bitburner-scripts/commit/d3521e7a0593776fa5aed2a09559a619b1ec2283))
+* **reclaim:** added server and network parts for reclaim stats ([c7fbd77](https://github.com/xjonsson/bitburner-scripts/commit/c7fbd77d5d0363f2ee590f451900c55570177cf6))
+* **scripts:** working on distribution and early game scripts ([02e2ebe](https://github.com/xjonsson/bitburner-scripts/commit/02e2ebe806d0e0b76e7b6ca2af3436d0767f566b))
+* **server:** added cores to Server class ([e255571](https://github.com/xjonsson/bitburner-scripts/commit/e255571d16f077286766c87b6f21c87611ba690b))
+* **server:** added functions for thread counts ([86d1bc0](https://github.com/xjonsson/bitburner-scripts/commit/86d1bc0d0df1e92e31f44ee7051afdb10db91a04))
 * **server:** added server class to help with network and calculations ([2bdc215](https://github.com/xjonsson/bitburner-scripts/commit/2bdc2152d94ae611dd392f91b97edfe24023efab))
 * **start:** added script to handle start and script selection ([f018c5a](https://github.com/xjonsson/bitburner-scripts/commit/f018c5af89294824a28b2d25d7d3d8bd22e57f76))
 * **udoor:** added utility script to backdoor all servers without singularity ([7777549](https://github.com/xjonsson/bitburner-scripts/commit/777754960e5d36062aec902947a88823e9a8ca1b))
@@ -24,4 +32,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Bug Fixes
 
+* **network:** added network ownership function for stats ([4b7607f](https://github.com/xjonsson/bitburner-scripts/commit/4b7607f528b93a31890389b37d7a01abf82779e9))
+* **reclaim:** fixed reclaim to work with server nodes ([449e14a](https://github.com/xjonsson/bitburner-scripts/commit/449e14a655657148f17d77f2a70bbeb9baebaa99))
 * **server:** removed redundant const from configs ([0af2513](https://github.com/xjonsson/bitburner-scripts/commit/0af25133f2c0f21f6001feb4ca1b372412838564))